requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Diffusion Models
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0

# Computer Vision
Pillow>=9.0.0
opencv-python>=4.7.0
scikit-image>=0.20.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
datasets>=2.12.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Progress & Logging
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
omegaconf>=2.3.0
hydra-core>=1.3.0
rich>=13.0.0
click>=8.1.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Optional: CUDA support
# torch-audio>=2.0.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
# torch-vision>=0.15.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
# torch>=2.0.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
